feat(essntl-3715): added tests
